gaussian fit
To run this code you need to work on a spectra which has peaks and dips. When you plot the spectra, than via the "np.ginput(2)" command you select two points on the graphs that are the left side and the right side of the peak (or dip) you want to fit. Inside the function only data inside this interval are going to be acquired.
The average of the second derivative of the data will allow the code to understand weather a peak or a dip has to be fitted, than taking first the logarithm and than a second order polyfit of the data you gonna work out parameters in order to find a proper fit. The thinks are not so easy when you encounter some negative or zero values which have imaginary values for the logarithm.
This code runs for my purposes and it could be used in order to fit other spectra